requirements.txt
flask==3.1.1
sqlalchemy==2.0.41
tushare==1.4.21
python-dotenv==1.1.1